+++ packages/express/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @lensjs/express
 
+## 1.3.2
+
+### Patch Changes
+
+- Fix Express Static Server
+
 ## 1.3.1
 
 ### Patch Changes